refactor(http): keep SSE policy in llm
